requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.6.0
backports.zoneinfo==0.2.1
backtrader==1.9.76.123
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
cycler==0.11.0
dateparser==1.1.1
Flask==2.1.1
fonttools==4.31.2
frozenlist==1.3.0
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
kiwisolver==1.4.2
MarkupSafe==2.1.1
matplotlib==3.2.2
multidict==6.0.2
numpy==1.22.3
packaging==21.3
Pillow==9.1.0
pycodestyle==2.8.0
pyparsing==3.0.7
python-binance==1.0.15
python-dateutil==2.8.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
regex==2022.3.2
requests==2.27.1
six==1.16.0
TA-Lib==0.4.24
toml==0.10.2
tzdata==2022.1
tzlocal==4.1
ujson==5.1.0
urllib3==1.26.9
websockets==9.1
Werkzeug==2.1.1
yarl==1.7.2
zipp==3.7.0
